+++ admin/class-parcelpro-admin.php
@@ -115,7 +115,10 @@ class Parcelpro_Admin
      */
     public function add_bulk_actions()
     {
-        if (get_current_screen()->id == 'woocommerce_page_wc-orders') {
+        global $post_type;
+
+        // Check for WooCommerce 7 and 8
+        if ($post_type == 'shop_order' || get_current_screen()->id == 'woocommerce_page_wc-orders') {
             include(plugin_dir_path(__FILE__) . 'partials/parcelpro-admin-actions-bulk.php');
         }
     }
@@ -270,11 +273,12 @@ class Parcelpro_Admin
                 wp_redirect($_SERVER['HTTP_REFERER'], 301);
                 exit;
             case 'parcelpro-bulk-export':
-                if (empty($_GET['id'])) {
+                // Check for WooCommerce 7 and 8
+                if (empty($_GET['post']) && empty($_GET['id'])) {
                     wp_die('Er zijn geen order geselecteerd!');
                 }
 
-                $order_ids = $_GET['id'];
+                $order_ids = empty($_GET['post']) ? $_GET['id'] : $_GET['post'];
 
                 foreach ($order_ids as $order_id) {
                     if (!$status = get_post_meta($order_id, '_parcelpro_status', true)) {
@@ -299,11 +303,12 @@ class Parcelpro_Admin
                 echo "<script>window.close();</script>";
                 exit;
             case 'parcelpro-bulk-label':
-                if (empty($_GET['id'])) {
+                // Check for WooCommerce 7 and 8
+                if (empty($_GET['post']) && empty($_GET['id'])) {
                     wp_die('Er zijn geen order geselecteerd!');
                 }
 
-                $order_ids = $_GET['id'];
+                $order_ids = empty($_GET['post']) ? $_GET['id'] : $_GET['post'];
                 $url = null;
 
                 foreach ($order_ids as $order_id) {